Documentation

Documentation

Environment

Features

I/O

DataHandling

Analysis

Extended Packages

Scattering Functions

Other

The package handles data parsing and analysis for inelastic neutron scattering experiments. It aims to be a flexible and reliable package for all needs for data ingestion and fast plotting, and holds routines for data analysis.

Environment

The package relies on many other Julia packages, and makes using them a bit easier by providing wrappers for specific functions.

Some of the most important Packages are:

DataFrames.jl and CSV.jl

Features

I/O

JuliaNeutronSpec.io_ill_header — Function

io_ill_header(filename::AbstractString)

Reads the ILL header and returns the three dictionaries

param with all the instrument configs

varia with all the motors

motor0 with all the zero values for the motors

df_meta with the line numbers important for reading the actual counts

JuliaNeutronSpec.io_ill — Function

io_ill(filename::AbstractString,
       precision_dict::AbstractDict{Symbol,Int};
       kwargs...)

Read a single experimental measurement scan from file. Handles the ILL format for inelastic neutron spectrometers (tested with IN20, IN8) in single detector and multidetector (flatcone) configuration. The output DataFrame has columns determined in columnsTAS. Motor angles can be overwritten by supplying a dictionary to the kwarg override.

Arguments

filename::AbstractString: relative file path

precision_dict::AbstractDict{Symbol,Int}: rounding to precision for defined column (precision_dict=Dict(:EN=>2)).

kwargs...:

rename_cols::Dict(): Dict(:old_name => :new_name)

ki::Number: only supply in ki=fixed measurements (time-of-flight)

kf::Number: manually supply kf

override::Dict(): Dict(:A1 = 15.0)

detector_efficiency::Array{Union{Float,Measurments}}

par_lat::Array{Float,1}: [a b c α β γ]

See also: columnsTAS, calc_detector_efficiency

JuliaNeutronSpec.calc_detector_efficiency — Method

calc_detector_efficiency(df)

calculate the detector efficiency for each detector distinguised by :detID using a single measurement. The output can be handed over to io_ill( ...; detector_efficiency=output), or the correction can be manually applied by dividing the Monitor count by the output.

Arguments

df::AbstractDataFrame: cols :CNTS and :detID are mandatory.

See also io_ill

DataHandling

JuliaNeutronSpec.combine — Function

combine(
    df_in::AbstractDataFrame,
    edges;
    bin_cols::Array{Symbol,1}=[:EN],
    monitor_count::Number=1e4
)

Combine multiple measurements to a single measurement. This function should only be applied to raw data!

All data in a DataFrame is binned according to edges along the columns defined in bin_cols and normalized to monitor_count.

JuliaNeutronSpec.normalize — Function

normalize!(
    df::AbstractDataFrame;
    monitor_col::Symbol=:MON,
    monitor_count::Float64=1e4,
    data_cols::Array{Symbol,1}=[:CNTS]
)

return a normalized DataFrame by scaling the column monitor_col to monitor_count, and applying the same scaling factor to the all data_cols as well.

See also: normalize

JuliaNeutronSpec.normalize! — Function

normalize!(
    df::AbstractDataFrame;
    monitor_col::Symbol=:MON,
    monitor_count::Float64=1e4,
    data_cols::Array{Symbol,1}=[:CNTS]
)

normalize a DataFrame in place by scaling the column monitor_col to monitor_count, and applying the same scaling factor to the all data_cols as well.

See also: normalize

Analysis

JuliaNeutronSpec.OrientationFactor — Function

orientationfactor_sq(x,y,z)

calcualte orientation factor for a,b,c direction from qx qy q_z in units of Å^-1

JuliaNeutronSpec.PolarizationAnalysis — Function

PolarizationAnalysis(df, MMON = 1e4, precision_dict)

converts NeutronDataFrames to polarizedNeutronDataFrames. sorts into SF / NSF, and x,y,z channels, and calculates polarization of excitations (relative to Q)

Extended Packages

Missing docstring.

Missing docstring for fit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Spline1D. Check Documenter's build log for details.

Scattering Functions

JuliaNeutronSpec.Kf — Function

Kf(kf, ϕ; ψ = 0)

calculate Kf from kf (Å^-1), ϕ (°), and ψ. ϕ is the horizontal detector angle. ψ is the vertical detector angle. TODO: implement out-of-plane scattering depending on ψ.

Other

JuliaNeutronSpec.columnsTAS — Constant

columnsTAS

OrderedDict that lists columns and corresponding etypes for DataFrames holding TAS Neutron Data.